% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caldav_get_all_simple_auth.R
\name{caldav_get_all_simple_auth}
\alias{caldav_get_all_simple_auth}
\title{Request data from CalDav Server}
\usage{
caldav_get_all_simple_auth(url, user = "", password = "")
}
\arguments{
\item{url}{url to caldav endpoint, e.g.: "https://nextcloud03.webo.cloud/remote.php/dav/calendars/me@myweb.de/personal/"}

\item{user}{user name}

\item{password}{user password}
}
\description{
Function to build and execute web request against CalDav server.
}
