% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ical_clean_ical_parsed.R
\name{ical_clean_ical_parsed}
\alias{ical_clean_ical_parsed}
\title{ical_clean_ical_parsed}
\usage{
ical_clean_ical_parsed(ical_parsed)
}
\arguments{
\item{ical_parsed}{data provided by \code{ical_parse()}}
}
\description{
ical_clean_ical_parsed
}
