% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ical_parse.R
\name{ical_parse}
\alias{ical_parse}
\title{ical_parse}
\usage{
ical_parse(file = NULL, text = NULL)
}
\arguments{
\item{file}{path to file to be read in and parsed}

\item{text}{text of ical file}
}
\description{
ical_parse
}
