% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ical_read_text.R
\name{ical_read_text}
\alias{ical_read_text}
\title{ical_read_text}
\usage{
ical_read_text(file = NULL, text = NULL, enc = "UTF-8")
}
\arguments{
\item{file}{a file path or connection object.}

\item{text}{a character vector.}

\item{enc}{encoding of the file to be read in}
}
\description{
Reads in text files or prepares character vectors.
Either \code{file} or \code{text} have to be specified.
}
