% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ical_test_files.R
\name{ical_test_files}
\alias{ical_test_files}
\title{ical_test_files}
\usage{
ical_test_files(fname = NULL)
}
\arguments{
\item{fname}{either NULL which will list files, an integer which will select file paths by index or
a string which will select files via regular expression}
}
\value{
returns one or more file paths
}
\description{
Function to get file paths to local test files included in the package.
}
\examples{

ical_test_files()
ical_test_files(1)
ical_test_files("mini")


}
