% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa_on_load.R
\docType{data}
\name{v8_env}
\alias{v8_env}
\title{env$v8 ...}
\format{
An object of class \code{environment} of length 1.
}
\usage{
v8_env
}
\description{
An environment containing bindings to an V8 instance for evaluating JS code.
}
\keyword{internal}
