% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ical_parse.R, R/ical_parse_df.R,
%   R/ical_parse_list.R
\name{ical_parse}
\alias{ical_parse}
\alias{ical_parse_df}
\alias{ical_parse_list}
\title{ical_parse}
\usage{
ical_parse(file = NULL, text = NULL)

ical_parse_df(file = NULL, text = NULL)

ical_parse_list(file = NULL, text = NULL)
}
\arguments{
\item{file}{path to file to be read in and parsed}

\item{text}{text of ical file}
}
\description{
Parses iCalendar format from character vector or file and returns it
as a list of columns.

Parses iCalendar format from character vector or file and returns it
as data.frame.

Parses iCalendar format from character vector or file and returns it
as a list of entries.
}
