% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ical_parse.R, R/ical_parse_df.R,
%   R/ical_parse_full.R, R/ical_parse_jcal.R, R/ical_parse_list.R
\name{ical_parse}
\alias{ical_parse}
\alias{ical_parse_df}
\alias{ical_parse_full}
\alias{ical_parse_jcal}
\alias{ical_parse_list}
\title{ical_parse}
\usage{
ical_parse(file = NULL, text = NULL, simplify = TRUE)

ical_parse_df(file = NULL, text = NULL)

ical_parse_full(file = NULL, text = NULL)

ical_parse_jcal(file = NULL, text = NULL)

ical_parse_list(file = NULL, text = NULL)
}
\arguments{
\item{file}{path to file to be read in and parsed}

\item{text}{text of ical file}

\item{return_empty_rows}{whether or not to filter out calendar items for
which all fields are empty (NA)}
}
\description{
Parses iCalendar format from character vector or file and returns it
as a list of columns.

Parses iCalendar format from character vector or file and returns it
as data.frame.

Parses all calendar data - to various formats: list, list of data.frames, object.

Parses iCalendar format from character vector or file and returns ALL data found
in the calendar as tree like list called jcal format.
This function is low level and expects the
programmer to figure out what belongs where and how to handle things.

Parses iCalendar format from character vector or file and returns it
as a list of entries.
}
\examples{

# parse from character vector
char_vec <- readLines(system.file("birthdays.ics", package = "ical"))
ical_parse(text = char_vec)

# parse from file
ical_parse(file = system.file("birthdays.ics", package = "ical"))

}
